adprism-ingest/requirements.txt
anthropic>=0.25.0
firebase-admin>=6.3.0
feedparser>=6.0.11
trafilatura>=1.8.0
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.0
pytz>=2024.1
google-genai>=1.0.0
